# chunk:  reduce chunk's iterator function call