<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
 Reduce chunk's iterator function call.

### What is changed and how it works?
Add `numRows int` to Iterator4Chunk.

Below is a benchmark
```go
func BenchmarkIterator4Chunk_Next(b *testing.B) {
        // 1024 rows
	chk := getChks()
	it := NewIterator4Chunk(chk)
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		for row := it.Begin(); row != it.End(); row = it.Next() {
			_ = row
		}
	}
}
```
old:
```
BenchmarkIterator4Chunk_Next     2000000              2712 ns/op
BenchmarkIterator4Chunk_Next     2000000              2711 ns/op
BenchmarkIterator4Chunk_Next     2000000              2704 ns/op
```
this branch 
```
BenchmarkIterator4Chunk_Next     3000000              1412 ns/op
BenchmarkIterator4Chunk_Next     3000000              1411 ns/op
BenchmarkIterator4Chunk_Next     3000000              1410 ns/op
```